refactor: renames navigationUtils into navigation
