+++ frontend/amundsen_application/static/js/components/ResourceListItem/DashboardListItem/index.tsx
@@ -8,7 +8,7 @@ import BookmarkIcon from 'components/Bookmark/BookmarkIcon';
 
 import { getSourceDisplayName, getSourceIconClass } from 'config/config-utils';
 import { logClick } from 'utils/analytics';
-import { buildDashboardURL } from 'utils/navigationUtils';
+import { buildDashboardURL } from 'utils/navigation';
 import { formatDate } from 'utils/date';
 
 import { ResourceType, DashboardResource } from 'interfaces';
