+++ frontend/amundsen_application/static/js/pages/DashboardPage/index.tsx
@@ -30,11 +30,7 @@ import {
   getResourceNotices,
 } from 'config/config-utils';
 import { formatDateTimeShort } from 'utils/date';
-import {
-  getLoggingParams,
-  getUrlParam,
-  setUrlParam,
-} from 'utils/navigationUtils';
+import { getLoggingParams, getUrlParam, setUrlParam } from 'utils/navigation';
 
 import { ResourceType } from 'interfaces';
 import { DashboardMetadata } from 'interfaces/Dashboard';
