+++ frontend/amundsen_application/static/js/pages/FeaturePage/index.tsx
@@ -39,11 +39,7 @@ import { PreviewDataTable } from 'features/PreviewData';
 import { FeatureMetadata, FeaturePreviewQueryParams } from 'interfaces/Feature';
 import { ResourceType } from 'interfaces/Resources';
 import { logAction } from 'utils/analytics';
-import {
-  getLoggingParams,
-  getUrlParam,
-  setUrlParam,
-} from 'utils/navigationUtils';
+import { getLoggingParams, getUrlParam, setUrlParam } from 'utils/navigation';
 import { formatDateTimeShort } from 'utils/date';
 import { ProgrammaticDescription } from 'interfaces';
 import EditableText from 'components/EditableText';
