+++ frontend/amundsen_application/static/js/pages/ProfilePage/index.spec.tsx
@@ -15,7 +15,7 @@ import globalState from 'fixtures/globalState';
 import { getMockRouterProps } from 'fixtures/mockRouter';
 import { ResourceType } from 'interfaces/Resources';
 
-import * as NavigationUtils from 'utils/navigationUtils';
+import * as NavigationUtils from 'utils/navigation';
 
 import { indexDashboardsEnabled } from 'config/config-utils';
 import { AVATAR_SIZE, PROFILE_TAB } from './constants';
