+++ frontend/amundsen_application/static/js/pages/TableDetailPage/LineageButton/index.tsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom';
 
 import { TableMetadata } from 'interfaces';
 import { isTableLineagePageEnabled } from 'config/config-utils';
-import { buildLineageURL } from 'utils/navigationUtils';
+import { buildLineageURL } from 'utils/navigation';
 
 export interface LineageButtonProps {
   tableData: TableMetadata;
