+++ frontend/amundsen_application/static/js/utils/navigation.ts
